Splitted postconditions in many files
