+++ lib/stones-spec.rb
@@ -11,7 +11,7 @@ require_relative 'stones-spec/with_gbb_html_rendering'
 require_relative 'stones-spec/version'
 
 require_relative 'stones-spec/precondition'
-require_relative 'stones-spec/postcondition'
+require_relative 'stones-spec/postcondition/postcondition'
 require_relative 'stones-spec/gobstones'
 require_relative 'stones-spec/example'
 require_relative 'stones-spec/subject'
